+++ stdlib/toolchain/CompatibilityDynamicReplacements/DynamicReplaceable.cpp
@@ -17,8 +17,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "swift/Runtime/Once.h"
 #include "swift/Runtime/Exclusivity.h"
+#include "swift/Runtime/FunctionReplacement.h"
+#include "swift/Runtime/Once.h"
 #include "swift/Runtime/ThreadLocalStorage.h"
 
 using namespace swift;
